Use async await for backend mutations

+++ src/resolvers/mutations/authors.js
@@ -1,40 +1,38 @@
-import { needGroup, ADMIN } from "../util/index";
+import { ADMIN, needGroup } from "../util/index";
 import { ObjectID } from "mongodb";
 
-export function addAuthor(root, { author }, context) {
+export async function addAuthor(root, { author }, context) {
+	await needGroup(context, ADMIN);
+
 	const id = new ObjectID();
-	return needGroup(context, ADMIN)
-		.then(() => Promise.resolve(author.picture))
-		.then(picture => picture && context.storage.save(id, picture))
-		.then(picture => {
-			author.picture = picture;
-			author._id = id;
-			return;
-			context.db
-				.collection("authors")
-				.insertOne(author)
-				.then(({ insertedId }) => insertedId);
-		});
+
+	if (author.picture) author.picture = await context.storage.save(id, picture);
+
+	author._id = id;
+
+	return context.db
+		.collection("authors")
+		.insertOne(author)
+		.then(({ insertedId }) => insertedId);
 }
 
-export function updateAuthor(root, { id, author }, context) {
-	return needGroup(context, ADMIN)
-		.then(() => Promise.resolve(author.picture))
-		.then(picture => picture && context.storage.save(id, picture))
-		.then(picture => {
-			author.picture = picture;
-			return context.db
-				.collection("authors")
-				.updateOne({ _id: new ObjectID(id) }, { $set: author })
-				.then(() => id);
-		});
+export async function updateAuthor(root, { id, author }, context) {
+	await needGroup(context, ADMIN);
+
+	if (author.picture)
+		author.picture = await context.storage.save(id, author.picture);
+
+	await context.db
+		.collection("authors")
+		.updateOne({ _id: new ObjectID(id) }, { $set: author });
+
+	return id;
 }
 
-export function deleteAuthor(root, { id }, context) {
-	return needGroup(context, ADMIN).then(() =>
-		context.db
-			.collection("authors")
-			.deleteOne({ _id: new ObjectID(id) })
-			.then(() => id)
-	);
+export async function deleteAuthor(root, { id }, context) {
+	await needGroup(context, ADMIN);
+
+	await context.db.collection("authors").deleteOne({ _id: new ObjectID(id) });
+
+	return id;
 }
